backend/pyproject.toml
# backend/pyproject.toml
[tool.poetry]
name = "requirements-manager"
version = "1.0.0"
description = "Requirements management system with FastAPI, LangGraph, and Celery"
authors = ["Engineering Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.9"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
redis = "^5.2.0"
celery = {extras = ["redis", "amqp"], version = "^5.4.0"}
langgraph = "^0.2.50"
structlog = "^24.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.27.2"
prometheus-client = "^0.21.0"
opentelemetry-api = "^1.28.2"
opentelemetry-sdk = "^1.28.2"
opentelemetry-instrumentation-fastapi = "^0.49b2"
opentelemetry-exporter-otlp = "^1.28.2"
opentelemetry-instrumentation-sqlalchemy = "^0.49b2"
opentelemetry-instrumentation-redis = "^0.49b2"
opentelemetry-instrumentation-celery = "^0.49b2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.2"
mypy = "^1.13.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
explicit_package_bases = true
no_implicit_optional = false
disable_error_code = ["var-annotated", "arg-type", "assignment", "index", "attr-defined", "union-attr", "no-any-return", "return-value", "typeddict-unknown-key"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --cov=app --cov-report=term-missing"

[tool.coverage.run]
omit = ["*/tests/*", "*/migrations/*", "*/__init__.py"]